Make nelder mead a separate module
